Remove useless garbage code
