--- audiofile/song_test.h
@@ -1,16 +1,0 @@
-#ifndef SONG_TEST_H
-#define SONG_TEST_H
-
-// audio file library tests
-void testCreateAudioFile();
-void testUpdateAudioFileByName(); 
-void testUpdateAudioFileById(); 
-void testDeleteAudioFile();
-void testDeleteAllAudioFiles();
-void testLoadAudioFiles();
-void testViewSongByName();
-void testViewSongById();  
-void testSyncAudioFilesPath(); 
-void runAudioFilesTest(); 
-
-#endif
