--- db/db_endpoints.h
@@ -1,9 +1,0 @@
-#ifndef DB_ENDPOINTS_H
-#define DB_ENDPOINTS_H
-
-#include "../endpoints/endpoints.h"
-
-void testDbFunctions(endpoint_t *e, char* argv[]);
-void dbEndpoints(endpoint_t **head); 
-
-#endif
