--- db/db_test.h
@@ -1,15 +1,0 @@
-#ifndef DB_TEST_H
-#define DB_TEST_H
-
-// database library tests
-void testOpenDb();
-void testCurrentTime();
-void testCombineFileStrs();
-void testRemoveChar();
-void testRenameFile();
-void testCountFilesDirectory();
-void testClearDirectory();
-void runDbTests(); 
-
-
-#endif
