+++ phaedra/phaedra_endpoints.c
@@ -1,6 +1,6 @@
 #include "phaedra.h"
 #include "phaedra_endpoints.h"
-#include "../db/db.h"
+#include "../logging/log.h"
 
 
 void playAudioCmd(endpoint_t *e, char* argv[]){
