--- playlist/playlist_test.h
@@ -1,13 +1,0 @@
-#ifndef PLAYLIST_TEST_H
-#define PLAYLIST_TEST_H
-
-// playlist library tests
-void testCreatePlaylistCollection();
-void testDeletePlaylistCollection();
-void testDeleteAllPlaylistCollections();
-void testDeleteCascadingPlaylists();
-void testLoadPlaylistsArray();
-void testAddSongToPlaylist(); 
-void runPlaylistTests(); 
-
-#endif
