--- urls/mp3_datasets/src/main.rs
@@ -1,13 +1,0 @@
-use youtube_dl::{YoutubeDl, YoutubeDlOutput};
-
-fn main() {
-    let output = YoutubeDl::new("https://www.youtube.com/watch?v=VFbhKZFzbzk")
-        .socket_timeout("15")
-        .run()
-        .unwrap();
-    let title = match output {
-        YoutubeDlOutput::SingleVideo(video) => video.title,
-        _ => panic!("single video should not be a playlist"),
-    };
-    println!("Video title: {}", title);
-}
