--- urls/urls_test.h
@@ -1,13 +1,0 @@
-#ifndef URLS_TEST_H
-#define URLS_TEST_H
-
-void testCreateUrl();
-void testDeleteUrl();
-void testDeleteAllUrls();
-void testLoadUrls();
-void testBuildArgumentThreads();
-void testSpawnArgumentThreads();
-void testGrabDatabaseUrls();
-void runUrlsTest(); 
-
-#endif
